[SERVICE-243] Moved interfaces used in tab events from (undocumented) internal.ts to tabbing.ts, removed some casting that is no longer necessary due to more robust type definitions.
